+++ Assets/CommonCore/Core/CoreParamsBase.cs
@@ -21,7 +21,7 @@ namespace CommonCore
         public static Version GameVersion { get; private set; } //auto-set from Unity settings
 
         //*****system version info
-        public static Version VersionCode { get; private set; } = new Version(6, 0, 0);
+        public static Version VersionCode { get; private set; } = new Version(7, 0, 0);
         public static string VersionName { get; private set; } = "Firene";
         public static Version UnityVersion { get; private set; } //auto-set
         public static string UnityVersionName { get; private set; } //auto-set